Float to string conversion using sprintf
I was experiencing problems with the library, namely due to an incorrect HTTP request (or incomplete JSON data).
What I found out was that the dtostrf() function wasn't working i.e. no conversion took place and str_val was esentially
an empty char array. The board I have used was a Tiva C TM4C123GXL with a CC3100 booster pack and my environment
was Energia IDE 1.8.7E21.

One way to circumvent around this was to use the standard sprintf() function with the %d format specifier as in the modified
source code.